+++ src/components/Data/Garten/index.js
@@ -102,7 +102,7 @@ const personQuery = gql`
   }
 `
 const personFelderQuery = gql`
-  query personFelderQuery($personId: Int) {
+  query personFelderQuery($personId: bigint) {
     person_felder(where: { person_id: { _eq: $personId } }) {
       ...PersonFelderFields
     }
