+++ src/components/Data/Teilkultur/Settings.js
@@ -51,7 +51,7 @@ const SettingsTeilkulturen = ({ teilkulturResult }) => {
         await client.mutate({
           mutation: gql`
               mutation update_kultur_felder(
-                $kulturId: Int!
+                $kulturId: bigint!
               ) {
                 update_kultur_felder(
                   where: { kultur_id: { _eq: $kulturId } }
