+++ src/components/Data/Zaehlung/Settings.js
@@ -52,7 +52,7 @@ const SettingsZaehlungen = ({ zaehlungResult }) => {
         await client.mutate({
           mutation: gql`
               mutation update_kultur_felder(
-                $kulturId: Int!
+                $kulturId: bigint!
               ) {
                 update_kultur_felder(
                   where: { kultur_id: { _eq: $kulturId } }
